+++ acceptance_tests/features/case_look_up.feature
@@ -2,17 +2,17 @@ Feature: Case look up for the contact centre
 
   Scenario: Find case by caseId
     Given sample file "sample_input_england_census_spec.csv" is loaded
-    Then the new cases are emitted to Respondent Home
+    Then messages are emitted to RH and Action Scheduler for with [01] qids
     And a case can be retrieved from the caseapi service
 
   Scenario: Find multiple cases from a single UPRN
     Given sample file "sample_input_england_census_spec.csv" is loaded
-    Then the new cases are emitted to Respondent Home
+    Then messages are emitted to RH and Action Scheduler for with [01] qids
     And caseapi returns multiple cases for a UPRN
 
   Scenario: Find case by caseRef
     When sample file "sample_input_england_census_spec.csv" is loaded
-    Then the new cases are emitted to Respondent Home
+    Then messages are emitted to RH and Action Scheduler for with [01] qids
     And a case can be retrieved by its caseRef
 
   Scenario: Check non-existent caseId returns a 404 status code
